Rename Mongo File to mongoSeed.js
